feat: enable auto scroll on menu
